requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
anyio==3.7.1
    # via starlette
bcrypt==4.0.1
    # via -r requirements.in
click==8.1.3
    # via uvicorn
exceptiongroup==1.1.2
    # via anyio
fastapi==0.99.1
    # via -r requirements.in
greenlet==2.0.2
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
idna==3.4
    # via anyio
mysql-connector-python==8.0.33
    # via -r requirements.in
protobuf==3.20.3
    # via mysql-connector-python
pydantic==1.10.11
    # via fastapi
python-dotenv==1.0.0
    # via -r requirements.in
sniffio==1.3.0
    # via anyio
sqlalchemy==2.0.18
    # via -r requirements.in
starlette==0.27.0
    # via fastapi
typing-extensions==4.7.1
    # via
    #   fastapi
    #   pydantic
    #   sqlalchemy
    #   starlette
uvicorn==0.22.0
    # via -r requirements.in
